Put classes in to seperate file. Looks cleaner now.
